refactor: restructure auth module
