+++ packages/auth/src/index.ts
@@ -1,7 +1,2 @@
-import axios from "axios";
-import { NavigraphCancelTokenStatic } from "./public-types";
-export * from "./api";
-export * from "./network";
-export * from "./public-types";
-
-export const CancelToken = axios.CancelToken as NavigraphCancelTokenStatic;
+export * from "./lib/navigraphRequest";
+export { default as getAuth } from "./lib/getAuth";
