+++ packages/auth/src/internals/loadPersistedCredentials.ts
@@ -1,0 +1,16 @@
+import { Logger } from "@navigraph/app";
+import verifyUser from "./verifyUser";
+import signOut from "./signOut";
+
+export let INITIALIZED = false;
+
+export default async function loadPersistedCredentials() {
+  if (INITIALIZED) return Promise.resolve();
+
+  await verifyUser().catch((e) => {
+    Logger.warning("Failed to load persisted credentials", e);
+    signOut().catch((e) => Logger.warning("Failed to sign out after failed initialization attempt", e));
+  });
+
+  INITIALIZED = true;
+}
