+++ packages/auth/src/lib/navigraphRequest.ts
@@ -1,10 +1,43 @@
 import { Logger, getApp } from "@navigraph/app";
 import axios, { AxiosError } from "axios";
-import { tokenCall } from "./flows/shared";
-import { tokenStorage } from "./internal";
-import { signOut } from "./internal";
+import { tokenStorage } from "../internals/storage";
+import { requestToken } from "../api/requestToken";
+import signOut from "../internals/signOut";
 
+export interface CancelStatic {
+  new (message?: string): Cancel;
+}
+
+export interface Cancel {
+  message: string;
+}
+
+export interface Canceler {
+  (message?: string): void;
+}
+
+export interface NavigraphCancelTokenStatic {
+  new (executor: (cancel: Canceler) => void): NavigraphCancelToken;
+  source(): CancelTokenSource;
+}
+
+/** A `CancelToken`, as defined by Axios. Can be used to abort ongoing sign-in attempts after polling has already started. */
+export interface NavigraphCancelToken {
+  promise: Promise<Cancel>;
+  reason?: Cancel;
+  throwIfRequested(): void;
+}
+
+export interface CancelTokenSource {
+  /** Cancellation token to be provided to the SDK */
+  token: NavigraphCancelToken;
+  /** Cancels any requests that are referencing the associated {@link NavigraphCancelToken} */
+  cancel: Canceler;
+}
+
+// Axios re-exports
 export const isAxiosError = (payload: unknown): payload is AxiosError => axios.isAxiosError(payload);
+export const CancelToken = axios.CancelToken as NavigraphCancelTokenStatic;
 
 export const navigraphRequest = axios.create();
 
@@ -28,7 +61,7 @@ navigraphRequest.interceptors.response.use(
     const REFRESH_TOKEN = await tokenStorage.getRefreshToken();
 
     if (app && error?.response?.status === 401 && REFRESH_TOKEN) {
-      const tokenResponse = await tokenCall({
+      const tokenResponse = await requestToken({
         client_id: app.clientId,
         client_secret: app.clientSecret,
         grant_type: "refresh_token",
